Set the working directory upon exit to orginal directory in SS_RunJitter
Users specify the directory to run the jitter iterations with the argument
mydir, which is used with setwd(mydir) at the start of SS_RunJitter, but
if this was not your working directory the function left the user in
mydir. This was rather annoying, and thus the commit changes the
working directory back to the working directory prior to calling
SS_RunJitter. It will change back even if the function fails.