Fixed stackpoly bug
This fixes a bug in stackpoly that would give an error when you tried to
make the plots with solid fill (instead of hashed). The ifelse()
functions that were used before would give this error when x.hash was
not specified: 'x' is NULL so the result will be NULL. This should fix
that problem and allow for making both solid and hashed filled polygons.