Add SSunavailableSpawningOutput plots to SS_plots, fix problems in SSunavailableSpawningOutput and SSplotData
In SS_plots, I added code to plot the SSunavailableSpawningOutput plots
at the the bottom of the selectivity tab.

In SSunavailableSpawningOutput, I corrected a few misspellings of the
word "spawning" (including in the function name) and changed the plot
info category to "Sel".

In the process of running the SS_plots function for several assessments,
I found an error in the SSplotData function where the variable "allyrs"
was misspelled as "allyears" so I fixed this bug.