avoid wrong warning.sso when jittering
Nancie Cummings notes that original output files are restored after
running SS_RunJitter with the exception of warning.sso, leading to
misleading messages if you read the output in this directory using
SS_output. Should be now fixed, and all warning files saved in case they
are useful in the future.