Add else statement when covar is changed to FALSE to not read head of covar
covar was already changed to FALSE, but the covar file was still being read in
within these brackets. Changed to only reading in the head of the covar file
if the file can be found, this will omit the time check. I cannot see any other
repurcussions from implementing this fix. No other code was changed, just indented
to match the new brackets.