Remove browser and commented code
Nothing was changed in the code base. Only calls to browser()
and a as.data.frame() that was commented out were removed.
The removal of "browser" is helpful when using findstr /s browser *.R
before commiting files, which decreases the chances of the unwanted
browser call ending up on github.