standard error of log catch is stored as one value per fleet
The value is overwritten for every year of catch data specific to
a given fleet. This is fine and the code works, but will not work
when there is more than one standard error for a given fleet. I don't
know if this would ever happen, but I wanted to include a check
so that the function that reads in the data tells the user that
the code should be changed.