fix: allow for matrix of Flimitfractions in forecast
If mylist$Flimitfraction is negative, a matrix of years and limits
will follow Flimitfraction. I use -999 to find the end of the matrix.
Then, many i values are skipped over so that the matrix values are not
wrongly put into other named values.

The write forecast function then puts writes these numbers back to the
file if Flimitfraction is negative.

I don't predict that the changes will have any conflicts with other
versions of SS.