fix: equilibrium catches were adding too many rows to catch
sum of the number of rows that equal -999 for year made multiple
rows in the new wide data frame, where there should only be one
additional row for equilibrium data.

Also fixed the standard error check that I made before which didn't
account for equilibrium catches maybe being in the data. Now it
removes the equilibrium catches to see if all standard errors
other than those are the same. @chantelwetzel-noaa, is this right?
Does the code work with different standard errors for year, did i wrongly
add this check in the first place?