fix: modification to read and write out the catch section for 3.30 models
This change rolls back some changes which had catches being formatted into a matrix form.  The matrix form prevented fleet specific se around catch.  Now catches are kept in 3.30 list form in the read and writing code.  This change requires users to use these functions across the same model version (read 3.30 and write 3.30). Users will not be able to read a 3.24 file and write it in 3.30 format.  If we think this is a needed option we will need to create some additional checks and switches.  However, since the transition executable will do this for users, I am not sure if we want to try to emulate this within r4ss.