Make parms for SS_readctl funs consistent
Some params missing in SS_readctl() that were present in the specific 3.24 and 3.30 versions of the functions, and SS_readctl_3.30 had an extra parameter DatFile that should be the same as datlist. Made parameter names consistent across the functions so they can work together (i.e., SS_readctl() can serve as a wrapper for the 3.24 and 3.30 functions).