Update reading MGpars by rec_dist_method
The method previously implemented addressed method 1, but not the other methods. This update will read the correct number of parameters for methods 2-4. It does not read 1, because this method is obsolete by 3.30.13 version of SS.

In the future, it may be desirable to incorporate back in the capability to read control files that use recr_dist_method = 1 for back compatibility.