fix: use convert.type in get.df for 3.30 data to get numeric values
get.df in SS_readdat_3.30 was returning character values
for numeric values. I first noticed this because I was trying to manipulate
some composition data. The CPUE data didn't display this issue because
SS_readdat converted some of those characters to numeric values.
I tried for a more general fix that would simplify the code.

convert.type is used by read.table to convert to logical, integer,
numeric, or complex and only if those fail will it generate character
values (or factors if as.is=TRUE)

I checked against two 3.30 data files that the results of the resulting
list were the same. I also went back in the history of the development
branch to ensure that the age and length comps had always been imported
as character values and that a recent change didn't create this problem.