fix: as.is = TRUE b/c 8b94b6e lead to vectors of characters being factors
Thanks @k-doering-NOAA for noticing that the fleetnames in the data list
was a vector of factors rather than a vector of character strings.

I ran this new fix with a dat file and the age comps are still numeric
while the fleetnames are now, again, a vector of characters rather than factors.