Initial commit SS_writectl and SS_writectl 3.30
Wrapper for SS_writectl_3.24 and SS_writectl_3.30 as well as first version for 3.30 (based heavily on already existing read/write control code).

Tested on the simple model and was able to read/write a control file and got the same model results with the generated control file as the orginal control file.