fix: change to endyr and depletion figure y-axis label
Modified the endyr to plot to endyr +1 where end year equals the model end year in the data file. Confirmed that this will generate plot through endyr + 1 for status determination plots and plots through endyr for F and SPR figures.  

Also, modified the y-axis for the depletion figure to "Fraction of unfished". The figure was labeled %unfished but was on a scale from 0-1 (typically).  This change makes it consistent to commit c69c94bd7e55451fc16ff27e05b1779952288b05 in the SSplotTimeseries function.