fix: unique() removed duplicated strings but could change intended order
goodnames creates a vector of strings based on the input strings.
I found a case where the parameter names were duplicated based on grep
then unique took the first instance, which led to the order of values in
newvals not matching the values in goodnames as I had intended.
Now, it checks to see if repeat.vals = FALSE and if there are repeated
names which then generates stop().