Development: read and write timevarying parlines, tests that functions work
Note that previous time varying code assumed that time varying was as in SS 3.24. It would be difficult to make this back compatible, so instead it was removed. Therefore, r4ss cannot be used to read in a 3.30 file and translate it back via writing using a 3.24 write function. This seemed acceptable.