feat(M_type): SS_ouput now reads M parameter and type
"Natural_Morality" is a section of the report file where the
natural mortality value is specified per age. This output is
potentially different than the temporal Z at age matrix where
there is no value for the terminal age category and the next to
last age category is an approximation.
I also read in the M_type now as a list element.
This is needed for Claudio Castillo-Jordan's work.