feature(SS_output): Use get_ncol to find optimum value of ncol in SS_output
Created an internal function (get_ncol) to determine the optimum width of the
table read in by SS_output of the Report.sso file. Previously, users had to
specify a value, where the default was 300. If this was too small a warning
would be generated and if it was too big the function was slow. Users can
now specify ncol=NULL and get_ncol will be called from SS_output to determine
the value to use. This number is printed to the screen if verbose = TRUE.

A test was added in test_simple to ensure the correct widths are being read in.
Also, added a test to see if a warning is generated when the mcmc folder
doesn't exist.

Use @template verbose for documentation and fixed a few typos in documentation.