fix: SS_readwtatage failed on blank lines
The hake assessment writes a weight at age file with blank lines
between the fleet numbers. The SS_readwtatage could not handle the
blank lines.

Now, blank lines are accommodated, the code is shorter, utilizes
read.table rather than read.table and readLines, the terminator line
is specific and no longer eliminates all negative year values which
are often used to fill in missing years, and a test was implemented.