fix: move removal of names in cpue to after check for cpue data
Commit a65b5e634114bd9884a59e27dd47073c2ed79b4d brought in a call to
remove some columns that were added in SS version 3.30.13. The removal
only works if there is CPUE data. For models without CPUE there will be
an error. The check is now moved to the else portion of the if else call
to ensure that stats$cpue doesn't just equal NA, which is what happens
when there is no CPUE info.