fix:init_yr is INIT in 3.30.13 for initial year F in "exploitation" matrix
init_yr is not used in 3.30.13, instead it is INIT. Use %in% to search for
either, making the initial year the year before the first year of the model.

Other changes include using type.convert to change all NA values including "_"
to NA in a single line of code rather than using a loop. Did this on the
matrix provided to returndat rather than the object b/c the object isn't
used again.