fix: make vectors the same length if NA values exist
Working with a model that had mean size at age data I found a small
warning message that the vectors were not the same length. I believe
with group 20 in SS_plots a vector of colours is passed rather than a
single value and when the NA values are subset from z the colour
vector was no longer the same length. I added the subsetting to the
colour vector that was already in place for x, y, and z.