use merge inside SSsummarize for indices
Comparing 2019 and 2020 models in bridging steps for Pacific Hake
and the summarize function was breaking down when trying to plot
the comparison of the indices. I have used merge elsewhere in
this function (likelihoods) when the column names didn't match up.
We should think about using merge more, though it does take longer
than rbind and may not be better in simple cases.