Bug fixes and naming uniformity
Corrected a few bugs in the 3.30 read control. Specifically changes to N_areas to Nareas to conform with read data and fix some null value reads. Also removed references to Nsurveys which are no longer a separate value in 3.30. Also converted Ngenders to Nsexes to conform with the read data function.