fix(SSplotData): Create space on plot for only kept data types
Found a bug in SSplotData when working with sablefish model because
each fleet does not have all data types. When SSplotData was fed
fleets = 1:2, then there were only catches, age compositions,
mean body weight, and discards but other fleets had abundance indices,
length compositions, and conditional age-at-length compositions.
So, the subset for fleets led to empty white space at the top of the
data plot b/c it was expecting seven data types where there were only
four. The fix creates the number of slots on the resulting figures
based on only the number of unique sample types still available
after the subset.

Added a full stop to the end of the caption that was missing.