fix(SS_readdat_3.30): MeanSize_at_Age_obs were not read if n on next line
Sometimes, the mean size at age sample sizes will be on a separate line,
which was leading to an error in SS_readdat_3.30. Now the sample sizes
are read in no matter what line they are on based. The code was tested
using ss3sim and files available in google drive where discard data
was added to the simple model, i.e., Simple with Discard.