fix(SS_output): lenntune had non-numeric values of _ for fleet 3
Copied code from agentune that changed _ to NA so type.convert can
change all values to numeric ones for mathematical operations.

Pushing to a fix branch to ensure travis doesn't fail because I am
not sure why it is currently passing on development. I found the error
when running check() on development branch. Once it passes a pull request
will be submitted.