refactor(SS_RunJitter): internal determination of OS & show.output.on.console=FALSE
* Mac and Linux were not being used except in else{} statement, so
I changed the if statement to directly determine if the OS is
Windows using .Platform rather than a multi-line call to grep and the
creation of an object.
* With shell or system, intern returns the output that is printed
in the command/bash window as an R object it does not technically supress
the output, so if you are trying to debug SS_RunJitter and intern=TRUE
and the call to system or shell is not saved to an R object, then
everything will be printed to the screen. Intern=TRUE only stops
the printing to the screen because it is called within SS_RunJitter().
On windows, show.output.on.console=FALSE, will supress the output.