refactor(SS_RunJitter): create file_increment to change file names
file_increment adds the appropriate integer to the .sso output files
and to the .par file, where for the .par file it becomes .par_i.sso
as it has always been changed. These names are normal for r4ss I just
removed duplicated code. We could also use this function in profile.