fix(SS_RunJitter): 2 errors
* setwd() was being called before check_model() and the path
being supplied to check_model could have been relative to the
original working directory. changed to use getwd() after setting
the working directory to wherever people want it to be seems to be
the best option going forward.
* 10 was being found with dir instead of just 0
Thanks @shcaba for the notification that once the jitter got to 10
there was an error printed to the screen. This was happening when
the original i.e., 0.sso files were being searched for to copy them
back over to the original files.