refactor, test: rm 3.24 var adjustment list from SS_readctl_3.30, make new fxn
This code is likely only helpful for a few users, and created 2 variance adjustment objects in the ctl list file which was confusing. It was better to move it to its own function so the functionality still exists for the few users that need it.

A test was also created.