#430 , feat: increase flexibility of exe location for run_SS_models
rm stop msg to allow any model path to be used. Note there are still some issues with using a model in your PATH vs not: even if mod_in_path = FALSE, this function may run a model in the PATH when trying to refer to a folder's model.

Note this also fixes the model run status, which sometimes reported success when the run actuallly failed.