doc(file): #311 use template and templateVar for file
@templateVar allows for brew to execute text based values of variables
defined in .R file. For example, I used if statements to determine which
type of file was specified to define the extra information. The default here
is to tell users to look at the formal arguments for a default filename.
Other types of files, such as posteriors, must be declared with
@templateVar file_t posteriors
and will result in extra text being added to the documentation. This was
helpful because the same verbage for posteriors is in multiple files.
This was necessary because @template calls cannot be nested :(