fix(SS_readctl) version was a vector for some files that aren't ss_new
Then substring() returned a vector and and error message for if(version %in%
was printed to the screen. Now, if it is a vector it is truncated to the
first element and if it is not a vector, as it shouldn't be, the function
still works because it still has an entry in the first element.