feat(SSplotComparisons): bring back bioscale feature for one-sex models
Was previously in r4ss, then removed, then added to hake2020 with
1c7a36d024688 . Basically cherry-picked this commit by hand because
hake2020 had gone stale. This is a clean branch as of 2021-01-16
that will be used for the hake2021 assessment and developed to
allow for the biomass to be scaled.

Can also change the control or data file, I cannot remember, to have
a value of -1 somewhere such that values are divided by 2.
Here we will use a bioscale of 0.5.