fix(SS_tune_comps): MI_mult for agecomps is wrong in SS 3.30.12-16
QUESTION: can someone check my proposed math in this commit to
ensure that I got the tuning correct?

recommended variance adjustment is wrong b/c it was set to 1.0.
@iantaylor-noaa knows and is fixing it in SS but r4ss will still
need to work with this output for 12-16 models.

Updated documentation based on ability to use markdown in the file,
which makes things easier to read, moved some sections.

Noticed during the 2021 pacific hake stock assessment