refactor(SS_tune_comps): define section = 1 for SS_readdat vs. NULL
NULL leads to section = 1 being defined in an if statement and then
a message printed to the screen. We don't need the message because
users will want the fitted values here not simulated data files.
The dat file is just used to help read in the control file, so providing
structure rather than actual values. By defining section = 1 the
message is no longer printed to the screen, which is helpful in
ss3sim where there are dat files with more sections and data weighting
is done internally.

Turn off message in SS_readdat_3.30 if section is defined as 1 and there
are more sections if verbose = FALSE. Didn't worry about previous
versions of SS_readdat.