refactor(SS_tune_comps): explicitly exclude covar if extras = "nohess"
extras = "nohess" will make it such that the covar file is not available
and SS_output will change covar to FALSE for you, but there is a warning
message that will be printed to the screen. By making it explicit, the
warning message is no longer printed to the screen.

I did a check with and without reading the covar
and the MI and Francis weightings were the
same, though I will note that the LO and HI Francis values in the table
were different based on reading the covar file or not. I don't think that
it matters for SS_tune_comps, but it could for other downstream things.
@ian-taylor-noaa were you aware of this behaviour?

also fixed one typo in a comment