feat(extract_sigma_R_info): pulled sigma_R_info code out of SS_output use tidyverse
Pulled the code out of SS_output to create extract_sigma_R_info. Changed it to use
tidyverse style coding. For the simple model this leads to taking 0.03 seconds
to create the data frame rather than 0.49 seconds. The savings for this one table
might not be much, but moving functionality out of SS_output into components could
save a lot of time. Also, this allows for calculating the same table using MCMC. A
feature that wasn't currently available and needed for hake 2021. Haven't actually
changed SS_output yet.