feat, #536: update option 2 der quant vecs
Inspired by changes made by KFJ, thanks Kelli!

Note that 3.30.17 will result in a breaking change to option 2, but this change provides a warning for users with 3.30.15 and 3.30.16 files. Unfortunately, could not find a way to make this fully back compatibile but provides the user with a warning and instructions.