fix(SS_writectl): check first if list and type in names
I ran into a problem where I got a non-sensible error message because
I fed SS_writectl() a list of objects rather than the ctllist.
So, now I first check to see if it is a list.
Second, I check to see if type is in the names.
Third, I allow SS_writectl to move on to the check that led to the error.