refactor(SS_profile): use argument `version` not `SSversion`
@iantaylor-NOAA you might not agree with all of the changes in this
commit :)

Changed SSversion to version in `SS_profile()`
Removed check for is.null(partfile) because hidden defaults
are frowned upon because users are not aware of what is going on.
Instead, here we basically are no longer providing support of SS_profile()
for model versions less than 3.30. If users want to use the function
they still can but they will need to modify parfile from the default
value.

Use the template to document version.

Part of #641