change to deprecate_stop if version=NULL
Also, implement last element to test if the input argument
of version does not equal what is stored in the ctllist.
This led to not being able to have NULL as an option
because the following if statement fails if version = NULL
because it has no length. So, deprecate_stop() was needed
instead of deprecate_warn().

Part of #641