fix(SS_writedat_3.30): Adds check for no mean size
Users might erroneously have `datlist[["use_MeanSize_at_Age_obs"]]`
be one when there are no observations. This happened to me in a
simulation when prior to sampling I had mean size, then
after sampling there were no years and all the data went away.
Thus, the new check ensures that the observations are not NULL
and if they are, then it sets the use value to 0.