Adds small details to documentation of copy_SS_inputs
Just refactoring the documentation is all. No major changes. When
reading I didn't understand what recursive did and I had to go to
the code to see where it was used which then led me to the documentation
for dir.create() so I thought I would just point users there in the
first place.