fix(SSmohnsrho): use endyr + 1 for biomass
based quantities because Stock Synthesis calculates beginning
of the year biomass thus endyr + 1 is the same as end of the year
biomass in the terminal year. This modification was done to
spawning biomass, recruitment, and fraction unfished because they
are all related to biomass. Updates to the documentation are also
included.