Fix retro to use endyr + 1 for biomass-based quantities (#824)
* doc(SSmohnsrho): style the documentation

* chore(SSmohnsrho): wrap at 80 characters, style

Fixed typos in warnings that are printed to the screen.

* fix(SSmohnsrho): use endyr + 1 for biomass

based quantities because Stock Synthesis calculates beginning
of the year biomass thus endyr + 1 is the same as end of the year
biomass in the terminal year. This modification was done to
spawning biomass, recruitment, and fraction unfished because they
are all related to biomass. Updates to the documentation are also
included.

* doc: run spelling::update_wordlist()

* add KFJ as co-author of SSmohnsrho()

---------

Co-authored-by: Ian Taylor <ian.taylor@noaa.gov>