fix(SSexecutivesummary): Works with all OS
specifying the file separator to be / can lead to difficulties on
various operating systems outside of Windows where `file.path()`
ensures that R is in charge of making appropriate file paths.
Rather than editing this twice for the two if statements, that should
have been an if else statement, I use `ifelse` to make the base
directory.