fix(SSexecutivesummary): Uses fleetnames vs.
replist[["FleetNames"]] when fleetnames exists. Removes the `names`
object that was not used but it was hard to tell that it was not being
used because names is also a function and a subset of many other
object names.

Now, the "default" option is also allowed, as is the case for some other
functions. Hopefully no one's fishery is named default :wink:

One of the replaced instances was not allowing for the user passed
fleetnames to be used and it was only getting information from the
names stored in replist[["FleetNames"]]