fix(SSexecutivesummary): Label landings as such
ES a had catch and dead, should be landings and dead catch but the
latter is just Dead in the header and dead catch in the caption.
Also total mortality did not make sense in the caption because it is
only fishing mortality.

Previously this was commit id 77d1901c where I also changed Dead to
Catch which was wrong and brought to my attention by @iantaylor-NOAA.
@chantelwetzel-noaa recommended just keeping dead even though it is
dead catch.