feat(SSexecutivesummary): Rm format, match_digits
Removes format and match_digits argument because the full un-rounded
numbers are saved in the csv file so you can round them however you
would like in your final output.

Only save .csv files because we as a team now largely rely on LaTeX
instead of needing to format things in microsoft Word. There will be
a function in sa4ss for that will create the .tex files for you from
the table of tables .csv file.

Also removed the functions that were present to facilitate rounding.