Fix SSexecutivesummary() (#889)
* fix(SSexecutivesummary): Uses _f.csv

natage and batage files can be sex specific but the automated file names
were using a full stop before f and an underscore before m, e.g.,
natage.f.csv and natage_m.csv.
This commit standardizes both files to use an underscore because the full
stop should be reserved for before the file extension.

* fix(SSexecutivesummary): Adds missing message()

For natage and batage there was a check and a string of text but no way
to get that text to the user through the console. They are now wrapped
in message and will be displayed only if `verbose = TRUE`.

* fix(SSexecutivesummary): Adds missing if("likes")

Likelihood table was always created even if not specified because there
was no wrapper if statement.

* fix(SSexecutivesummary): Removes `stop()` check

if (is.null(replist)) ... after talking with @iantaylor-NOAA he okayed
this removal.

Also, removes old comment about file needing to be in a certain
directory that does not pertain to the current code, did not pertain
before any of the changes that I made.

* fix(SSexecutivesummary): Works with all OS

specifying the file separator to be / can lead to difficulties on
various operating systems outside of Windows where `file.path()`
ensures that R is in charge of making appropriate file paths.
Rather than editing this twice for the two if statements, that should
have been an if else statement, I use `ifelse` to make the base
directory.

* doc(SSexecutivesummary): Modernize documentation

* fix(SSexecutivesummary): Removes usused object

* fix(SSexecutivesummary): Uses fleetnames vs.

replist[["FleetNames"]] when fleetnames exists. Removes the `names`
object that was not used but it was hard to tell that it was not being
used because names is also a function and a subset of many other
object names.

Now, the "default" option is also allowed, as is the case for some other
functions. Hopefully no one's fishery is named default :wink:

One of the replaced instances was not allowing for the user passed
fleetnames to be used and it was only getting information from the
names stored in replist[["FleetNames"]]

* fix(SSexecutivesummary): Label landings as such

ES a had catch and dead, should be landings and dead catch but the
latter is just Dead in the header and dead catch in the caption.
Also total mortality did not make sense in the caption because it is
only fishing mortality.

Previously this was commit id 77d1901c where I also changed Dead to
Catch which was wrong and brought to my attention by @iantaylor-NOAA.
@chantelwetzel-noaa recommended just keeping dead even though it is
dead catch.

* fix(SSexecutivesummary): decimal fraction unfished

Stop multiplying fraction unfished in time series table by 100, which
means that the digits had to be changed from 1 to 3. Fraction unfished
is not a percent, otherwise it would be percent unfished.

* fix(SSexecutivesummary): Adds colname to first col of table e

Reference Points ES table used row names for the first column, which
did not allow for the column to be named. Now the old row names are the
first column, which is named. Thus, row names had to be omitted from
the writing of the csv.

* fix(SSexecutivesummary): year before label table e

2023 Spawning Biomass
2023 Fraction Unfished
instead of
Spawning Biomass (mt) (2023)
Fraction Unfished (2023)
because year is not really a unit and double parentheses are :frown:

* chore(WORDLIST): adds the following words

formatting
microsoft
overfishing
unformatted

* feat(SSexecutivesummary): Rm format, match_digits

Removes format and match_digits argument because the full un-rounded
numbers are saved in the csv file so you can round them however you
would like in your final output.

Only save .csv files because we as a team now largely rely on LaTeX
instead of needing to format things in microsoft Word. There will be
a function in sa4ss for that will create the .tex files for you from
the table of tables .csv file.

Also removed the functions that were present to facilitate rounding.

* refactor(SSexecutivesummary): fraction_unfished

instead of depl in all instances. Also, changed the comments and
messages to Fraction Unfished. No changes to output.

* fix(SSexecutivesummary): Deprecate tables argument

Users can no longer choose which table they would like to create.
Instead, all tables are created which is the same behavior as the
previous default behavior. Users will be warned if they try to limit
the tables that are created but the function will still work.

* doc(SSexecutivesummary): Ran devtools::document()

* refactor(SSexecutivesummary): Wraps some lines at 80

* add authors to SSexecutivesummary.r

---------

Co-authored-by: Ian Taylor <ian.taylor@noaa.gov>